Inline AbstractArrayBlock
