+++ plugin/trino-thrift-api/src/main/java/io/trino/plugin/thrift/api/datatypes/TrinoThriftBigintArray.java
@@ -17,7 +17,6 @@ import io.airlift.drift.annotations.ThriftConstructor;
 import io.airlift.drift.annotations.ThriftField;
 import io.airlift.drift.annotations.ThriftStruct;
 import io.trino.plugin.thrift.api.TrinoThriftBlock;
-import io.trino.spi.block.AbstractArrayBlock;
 import io.trino.spi.block.ArrayBlock;
 import io.trino.spi.block.Block;
 import io.trino.spi.block.LongArrayBlock;
@@ -153,8 +152,8 @@ public final class TrinoThriftBigintArray
             Arrays.fill(nulls, true);
             return bigintArrayData(new TrinoThriftBigintArray(nulls, null, null));
         }
-        checkArgument(block instanceof AbstractArrayBlock, "block is not of an array type");
-        AbstractArrayBlock arrayBlock = (AbstractArrayBlock) block;
+        checkArgument(block instanceof ArrayBlock, "block is not of an array type");
+        ArrayBlock arrayBlock = (ArrayBlock) block;
 
         boolean[] nulls = null;
         int[] sizes = null;
